refactor: move ticket update out of ActionRemark
